requirements.txt
beautifulsoup4==4.12.2
fastapi==0.95.1
mysql_connector_repackaged==0.3.1
pandas==1.4.2
Pillow==9.5.0
pytesseract==0.3.9
python-dotenv==1.0.0
requests==2.25.1
selenium==4.8.3
SQLAlchemy==1.4.36
uvicorn==0.21.1
